🚚 move files according to #652
