+++ src/visitors/pages/SignIn/QrScanner.js
