+++ src/visitors/pages/SignIn/SignInActivitiesGridItem.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { PrimaryButtonNoFill, SecondaryButton } from '../../shared/components/form';
+import { PrimaryButtonNoFill, SecondaryButton } from '../../../shared/components/form';
 
 
 const BtnOne = styled(PrimaryButtonNoFill)`
